Change ResolutionScope for the java:java goal
The ExecJavaMojo is using ResolutionScope.TEST which triggers a
dependency resolution error when running the goal in the compile phase
and a test-jar dependency is present in the project.

The workaround was to run "mvn test-compile" instead of "mvn compile" to
ensure that the test-jar dependency can be resolved.

This change updates the ExecJavaMojo to use ResolutionScope.COMPILE
instead of ResolutionScope.TEST to avoid the test-jar resolution error.

See: https://github.com/mojohaus/exec-maven-plugin/issues/106